+++ AccountingPolessUp/Views/Administration/PageAdmProjects.xaml.cs
@@ -79,6 +79,7 @@ namespace AccountingPolessUp.Views.Administration
         }
         private void ButtonDelete_Click(object sender, RoutedEventArgs e)
         {
+           
             DeleteSelectedProjects();
         }
         private void Number_PreviewTextInput(object sender, TextCompositionEventArgs e)
@@ -95,7 +96,8 @@ namespace AccountingPolessUp.Views.Administration
             {
                 foreach (Project project in dataGrid.SelectedItems)
                 {
-                    _projectService.Delete(project.Id);
+                    if (RoleValidator.RoleChecker((int)project.idLocalPM))
+                        _projectService.Delete(project.Id);
                 }
             }
             UpdateDataGrid();
@@ -112,8 +114,11 @@ namespace AccountingPolessUp.Views.Administration
         {
             foreach (Project project in dataGrid.SelectedItems)
             {
-                EditFrame.Content = new PageEditProject(project, this);
-                break;
+                if (RoleValidator.RoleChecker((int)project.idLocalPM))
+                {
+                    EditFrame.Content = new PageEditProject(project, this);
+                    break;
+                }
             }
         }
         private void ButtonStages_Click(object sender, RoutedEventArgs e)
